new classes queues organization and double key for heap

+++ utils/priority_queue/heap/_heap.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 
 typedef struct _heap_entry{
-	int key;
+	double key;
     void * payload;
 	int position;
 } _heap_entry;
