+++ utils/priority_queue/heap/heap.h
@@ -16,7 +16,7 @@ heap * heap_new(HEAP_TYPE type, int capacity);
 HEAP_TYPE heap_type(heap * h);
 int heap_peek(heap * h);
 void * heap_poll(heap * h);
-heap_entry * heap_add(heap * h, int key, void * payload);
+heap_entry * heap_add(heap * h, double key, void * payload);
 int get_key_entry(heap_entry * e);
 int heap_size(heap * h);
 void heap_delete(heap * h);
@@ -26,7 +26,7 @@ heap * array2heap(int * array, int size, HEAP_TYPE type);
 
 void heap_sort(int * array, int size);
 
-void heap_update_key(heap * h, heap_entry * e, int key);
+void heap_update_key(heap * h, heap_entry * e, double key);
 
 #ifdef	__cplusplus
 }
